Update db data and schema

+++ src/main/resources/db/h2/schema.sql
@@ -24,7 +24,7 @@ CREATE TABLE IF NOT EXISTS education_center (
 	name VARCHAR,
 	city_id INTEGER REFERENCES city(id),
 	registration_fee VARCHAR,
-	rating INTEGER
+	rating REAL
 );
 
 CREATE TABLE IF NOT EXISTS course (
